refac: 에러 페이지 모듈화
